Added implementation file for Ghost class
